refactor(ui): extract TaskDetail pure helpers + AgentConversation
TaskDetail.tsx held 1681 LOC of mixed React + activity logic + a
sizable conversation cluster. Split into three modules per-concern;
the React route component stays at the top.

  TaskDetail.tsx:           1681 → 1064 LOC (-617 in route component)
  taskDetailHelpers.ts:     new, 302 LOC (pure helpers)
  TaskAgentConversation:    new, 357 LOC (conversation + retry modal)
  Net: +42 LOC for import boilerplate

What moves:

taskDetailHelpers.ts — pure functions, no React:
  - STEP_STATUS_COLOR, stepColor, splitNamespacedToolName,
    taskBadgeStatus, approvalCommandPreview, describeRunEvent,
    isVisibleRunEvent, describeRunEventNote, describeApprovalKind
  - OutputActivityIndex + buildOutputActivityIndex,
    failedToolOutputArtifacts, isOutputArtifactActivity,
    outputActivityStream
  - taskActivityAdvancedRows, taskActivityToTranscriptActivity,
    taskActivityArtifactSize, taskActivityArtifactPreview,
    artifactHasBytes, taskActivityTitle, taskActivitySubtitle,
    summaryString, summaryNumber, nonInternalKind,
    isVisibleArtifactBadge

TaskAgentConversation.tsx — AgentConversationView + RetryFromTurnModal
+ ConversationBubble + ToolCallChip components. Exports
AgentConversationView, isVisibleConversationMessage,
buildTurnCostMap; internal types and bubble/chip stay module-private.

What stays in TaskDetail.tsx: the route component, the run header,
the run timeline, the step inspector, the artifact list — the
pieces that read AppShell context and own page layout.

External callsites unchanged. No production behavior change.
