+++ ui/src/features/runs/TaskDetail.tsx
@@ -1,57 +1,35 @@
 import { useEffect, useMemo, useRef, useState, type ReactNode } from "react";
-import type { AgentChatActivityRecord, TaskActivityRecord, TaskApprovalRecord, TaskArtifactRecord, TaskRecord, TaskRunEventRecord, TaskRunRecord, TaskStepRecord } from "../../types/runtime";
+import type { TaskActivityRecord, TaskApprovalRecord, TaskArtifactRecord, TaskRecord, TaskRunEventRecord, TaskRunRecord, TaskStepRecord } from "../../types/runtime";
 import { formatDurationRange, formatLocaleDateTime, formatLocaleTime, formatMicrosUSD } from "../../lib/format";
 import { Badge, BrandAvatar, Dot, Icon, Icons, Modal } from "../shared/ui";
 import { TranscriptActivityTimeline } from "../transcript/TranscriptActivityTimeline";
 
+import { AgentConversationView } from "./TaskAgentConversation";
+import {
+  type OutputActivityIndex,
+  approvalCommandPreview,
+  artifactHasBytes,
+  buildOutputActivityIndex,
+  describeApprovalKind,
+  describeRunEvent,
+  describeRunEventNote,
+  failedToolOutputArtifacts,
+  isOutputArtifactActivity,
+  isVisibleArtifactBadge,
+  isVisibleRunEvent,
+  outputActivityStream,
+  splitNamespacedToolName,
+  stepColor,
+  summaryNumber,
+  summaryString,
+  taskActivityAdvancedRows,
+  taskActivityArtifactPreview,
+  taskActivityArtifactSize,
+  taskActivityToTranscriptActivity,
+  taskBadgeStatus,
+} from "./taskDetailHelpers";
+
 type StreamState = "idle" | "connecting" | "live" | "closed" | "error";
-
-const STEP_STATUS_COLOR: Record<string, string> = {
-  completed: "var(--green)",
-  running:   "var(--teal)",
-  awaiting_approval: "var(--amber)",
-  failed:    "var(--red)",
-  cancelled: "var(--red)",
-};
-function stepColor(status: string) { return STEP_STATUS_COLOR[status] || "var(--t3)"; }
-
-// MCP_TOOL_PREFIX mirrors mcpclient.PoolToolNamespacePrefix +
-// PoolToolNamespaceSep on the gateway side. Tool calls dispatched
-// through an external MCP server arrive here as steps whose
-// tool_name follows this shape:
-//
-//   mcp__<server>__<tool>
-//
-// We split it so the timeline can render server + tool separately
-// (operators want to see "github · create_pr", not the raw
-// double-underscore string), and the StepDetail can break out
-// transport / server / tool labels.
-const MCP_TOOL_PREFIX = "mcp__";
-const MCP_TOOL_SEP = "__";
-
-// splitNamespacedToolName mirrors the Go-side SplitNamespacedToolName
-// (internal/mcp/client/pool.go). Returns server + tool when the name
-// matches the namespacing scheme, otherwise null. Tool names may
-// themselves contain "__" — we honor only the FIRST split after the
-// server segment, so `mcp__weird__double__under` parses as
-// (weird, double__under), matching the gateway's split.
-type MCPToolName = { server: string; tool: string };
-function splitNamespacedToolName(name: string | undefined): MCPToolName | null {
-  if (!name || !name.startsWith(MCP_TOOL_PREFIX)) return null;
-  const rest = name.slice(MCP_TOOL_PREFIX.length);
-  const idx = rest.indexOf(MCP_TOOL_SEP);
-  if (idx <= 0) return null;
-  const server = rest.slice(0, idx);
-  const tool = rest.slice(idx + MCP_TOOL_SEP.length);
-  if (!server || !tool) return null;
-  return { server, tool };
-}
-
-function taskBadgeStatus(status: string): string {
-  if (status === "completed") return "done";
-  if (status === "awaiting_approval") return "awaiting";
-  return status;
-}
 
 // RunCostBadge shows this run's cost — and, when prior runs exist
 // in the resume chain, the task-cumulative figure too. Operators
@@ -252,75 +230,6 @@ function TaskApprovalCallout({
       ))}
     </div>
   );
-}
-
-function approvalCommandPreview(task: TaskRecord): string {
-  if (task.execution_kind === "git" && task.git_command) return `git ${task.git_command}`;
-  if (task.shell_command) return task.shell_command;
-  if (task.file_path) return `${task.file_operation || "write"} ${task.file_path}`;
-  return "";
-}
-
-function describeRunEvent(eventType: string): { label: string; tone: "queued" | "running" | "awaiting" | "done" | "failed" } {
-  const labels: Record<string, { label: string; tone: "queued" | "running" | "awaiting" | "done" | "failed" }> = {
-    "run.created": { label: "Run created", tone: "queued" },
-    "run.queued": { label: "Queued", tone: "queued" },
-    "run.started": { label: "Started", tone: "running" },
-    "run.awaiting_approval": { label: "Approval wait", tone: "awaiting" },
-    "run.cancelled": { label: "Cancelled", tone: "failed" },
-    "run.failed": { label: "Failed", tone: "failed" },
-    "run.finished": { label: "Completed", tone: "done" },
-    "run.resumed_from_event": { label: "Resumed", tone: "running" },
-    "gap.run_disconnected": { label: "Runtime recovered", tone: "queued" },
-    "turn.started": { label: "Turn started", tone: "running" },
-    "turn.completed": { label: "Turn done", tone: "done" },
-    "assistant.tool_call_proposed": { label: "Tool proposed", tone: "queued" },
-    "tool.invoked": { label: "Tool invoked", tone: "running" },
-    "tool.started": { label: "Tool started", tone: "running" },
-    "tool.shell_command": { label: "Shell command", tone: "running" },
-    "tool.failed": { label: "Tool failed", tone: "failed" },
-    "tool.completed": { label: "Tool done", tone: "done" },
-    "approval.requested": { label: "Approval asked", tone: "awaiting" },
-    "approval.resolved": { label: "Approval done", tone: "done" },
-  };
-  return labels[eventType] ?? { label: eventType.replaceAll("_", " "), tone: "queued" };
-}
-
-function isVisibleRunEvent(event: TaskRunEventRecord): boolean {
-  return event.type !== "snapshot" && event.type !== "run.snapshot";
-}
-
-// describeRunEventNote extracts a human-readable annotation from an event's
-// data payload. Returns null when there is nothing worth surfacing.
-//
-// Covers two axes:
-//   retry_from_turn — the turn number a retry-from-turn was branched at
-//   reason          — the operator's annotation for why they resumed/branched
-//
-// run.resumed_from_event stores the operator annotation under "reason".
-function describeRunEventNote(event: { data?: Record<string, unknown> }): string | null {
-  const d = event.data;
-  if (!d) return null;
-  const turn = typeof d["retry_from_turn"] === "number" ? (d["retry_from_turn"] as number) : null;
-  const reason = (
-    typeof d["reason"] === "string" ? d["reason"] : ""
-  ).trim();
-  if (!turn && !reason) return null;
-  const parts: string[] = [];
-  if (turn) parts.push(`turn ${turn}`);
-  if (reason) parts.push(reason);
-  return parts.join(" — ");
-}
-
-function describeApprovalKind(kind: string): string {
-  switch (kind) {
-    case "shell_command":        return "Shell execution";
-    case "git_exec":             return "Git execution";
-    case "file_write":           return "File write";
-    case "network_egress":       return "Network egress";
-    case "agent_loop_tool_call": return "Agent tool call";
-    default:                     return kind.replaceAll("_", " ");
-  }
 }
 
 type Props = {
@@ -885,11 +794,6 @@ function RuntimeActivity({ activity }: { activity: TaskActivityRecord[] }) {
   );
 }
 
-type OutputActivityIndex = {
-  all: TaskActivityRecord[];
-  byStepID: Map<string, TaskActivityRecord[]>;
-};
-
 function TaskActivityAdvancedDetails({ activity, outputArtifacts }: { activity?: TaskActivityRecord; outputArtifacts: OutputActivityIndex }) {
   if (!activity) return null;
   const rows = taskActivityAdvancedRows(activity);
@@ -1081,168 +985,6 @@ function TaskActivityOutputPreview({ artifact }: { artifact: TaskActivityRecord 
   );
 }
 
-function buildOutputActivityIndex(activity: TaskActivityRecord[]): OutputActivityIndex {
-  const all: TaskActivityRecord[] = [];
-  const byStepID = new Map<string, TaskActivityRecord[]>();
-  const seen = new Set<string>();
-  for (const item of activity) {
-    if (item.type !== "artifact" || !isOutputArtifactActivity(item)) continue;
-    const key = item.artifact_id || item.id;
-    if (seen.has(key)) continue;
-    seen.add(key);
-    all.push(item);
-    if (item.step_id) {
-      const scoped = byStepID.get(item.step_id) ?? [];
-      scoped.push(item);
-      byStepID.set(item.step_id, scoped);
-    }
-  }
-  return { all, byStepID };
-}
-
-function failedToolOutputArtifacts(activity: TaskActivityRecord, outputArtifacts: OutputActivityIndex): TaskActivityRecord[] {
-  if (activity.type !== "tool_call" || activity.status !== "failed") return [];
-  const matchingStep = activity.step_id || "";
-  if (matchingStep) {
-    return outputArtifacts.byStepID.get(matchingStep) ?? [];
-  }
-  return outputArtifacts.all;
-}
-
-function isOutputArtifactActivity(activity: TaskActivityRecord): boolean {
-  return outputActivityStream(activity) !== "";
-}
-
-function outputActivityStream(activity: TaskActivityRecord): "stdout" | "stderr" | "" {
-  const label = `${activity.kind ?? ""} ${activity.title ?? ""} ${activity.path ?? ""}`.toLowerCase();
-  if (label.includes("stderr")) return "stderr";
-  if (label.includes("stdout")) return "stdout";
-  return "";
-}
-
-function taskActivityAdvancedRows(activity: TaskActivityRecord): Array<{ label: string; value: string; multiline?: boolean }> {
-  const rows: Array<{ label: string; value: string; multiline?: boolean }> = [
-    ["type", activity.type],
-    ["status", activity.status],
-    ["occurred", activity.occurred_at],
-    ["activity", activity.id],
-    ["step", activity.step_id],
-    ["artifact", activity.artifact_id],
-    ["approval", activity.approval_id],
-    ["tool", activity.tool_name],
-    ["kind", activity.kind],
-    ["path", activity.path],
-    ["needs action", activity.needs_action ? "yes" : ""],
-    ["terminal", activity.terminal ? "yes" : ""],
-  ]
-    .filter((row): row is [string, string] => Boolean(row[1]))
-    .map(([label, value]) => ({ label, value }));
-
-  if (activity.summary && Object.keys(activity.summary).length > 0) {
-    rows.push({
-      label: "summary",
-      value: JSON.stringify(activity.summary, null, 2),
-      multiline: true,
-    });
-  }
-
-  return rows;
-}
-
-function taskActivityToTranscriptActivity(item: TaskActivityRecord): AgentChatActivityRecord {
-  return {
-    id: item.id,
-    type: item.type,
-    status: item.needs_action ? "awaiting_approval" : item.status,
-    title: taskActivityTitle(item),
-    kind: item.kind,
-    detail: taskActivitySubtitle(item),
-    created_at: item.occurred_at,
-    artifact_id: item.artifact_id,
-    artifact_size_bytes: taskActivityArtifactSize(item),
-    artifact_preview: taskActivityArtifactPreview(item),
-    approval_id: item.approval_id,
-    needs_action: item.needs_action,
-    terminal: item.terminal,
-  };
-}
-
-function taskActivityArtifactSize(item: TaskActivityRecord): number | undefined {
-  const value = item.summary?.size_bytes;
-  return typeof value === "number" ? value : undefined;
-}
-
-function taskActivityArtifactPreview(item: TaskActivityRecord): string | undefined {
-  const value = item.summary?.content_preview;
-  return typeof value === "string" && value.trimEnd() ? value.replace(/[\r\n]+$/, "") : undefined;
-}
-
-function artifactHasBytes(artifact: TaskArtifactRecord): boolean {
-  if (typeof artifact.size_bytes === "number") return artifact.size_bytes > 0;
-  return Boolean(artifact.content_text);
-}
-
-function taskActivityTitle(item: TaskActivityRecord): string {
-  switch (item.type) {
-    case "approval":
-      if (item.needs_action || item.status === "pending" || item.status === "awaiting_approval") return "Waiting for approval";
-      if (item.status === "approved") return "Approval granted";
-      if (item.status === "rejected" || item.status === "denied") return "Approval rejected";
-      return "Approval";
-    case "artifact":
-      return "Artifact";
-    case "changed_files":
-      return "Changed files";
-    case "final_answer":
-      return "Final answer artifact";
-    case "patch":
-      return "Patch";
-    case "tool_call":
-      return item.tool_name || item.title || item.path || "tool";
-    default:
-      return item.title || item.tool_name || item.path || item.type.replaceAll("_", " ");
-  }
-}
-
-function taskActivitySubtitle(item: TaskActivityRecord): string | undefined {
-  const status = item.status || "";
-  const reason = summaryString(item, "reason");
-  const command = summaryString(item, "command");
-  const filename = item.path || item.title || "";
-  const parts = (() => {
-    switch (item.type) {
-      case "approval":
-        return [reason, nonInternalKind(item.kind)];
-      case "tool_call":
-        return [item.path, command];
-      case "artifact":
-      case "changed_files":
-      case "final_answer":
-        return [filename, status && status !== "ready" ? status : ""];
-      case "patch":
-        return [filename, status && status !== "ready" ? status : ""];
-      default:
-        return [item.path, nonInternalKind(item.kind), status];
-    }
-  })().filter(Boolean);
-  return parts.join(" · ") || undefined;
-}
-
-function summaryString(item: TaskActivityRecord, key: string): string {
-  const value = item.summary?.[key];
-  return typeof value === "string" ? value.trim() : "";
-}
-
-function summaryNumber(item: TaskActivityRecord, key: string): number | undefined {
-  const value = item.summary?.[key];
-  return typeof value === "number" && Number.isFinite(value) ? value : undefined;
-}
-
-function nonInternalKind(kind?: string): string {
-  const value = kind?.trim() || "";
-  return value.startsWith("builtin.agent_loop_") ? "" : value;
-}
-
 function StepDetail({ step }: { step: TaskStepRecord }) {
   const duration = formatDurationRange(step.started_at, step.finished_at);
   const mcp = splitNamespacedToolName(step.tool_name);
@@ -1320,362 +1062,3 @@ function StepDetail({ step }: { step: TaskStepRecord }) {
   );
 }
 
-// isVisibleArtifactBadge filters which artifacts get a chip in the
-// bottom strip. stdout/stderr are rendered in the terminal pane;
-// agent_conversation is rendered as a chat-bubble timeline. Both
-// would be redundant as bare chips, so we hide them.
-function isVisibleArtifactBadge(a: TaskArtifactRecord): boolean {
-  return a.kind !== "stdout" && a.kind !== "stderr" && a.kind !== "agent_conversation";
-}
-
-// AgentConversationMessage mirrors pkg/types.Message (the shape the
-// agent loop persists). Only fields the viewer renders are typed —
-// extra fields on the wire (cache control, etc.) are ignored.
-type AgentConversationMessage = {
-  role: "user" | "assistant" | "tool" | string;
-  content?: string;
-  tool_call_id?: string;
-  tool_calls?: Array<{
-    id: string;
-    type?: string;
-    function?: { name?: string; arguments?: string };
-  }>;
-};
-
-// AgentConversationView renders the agent_loop conversation as a
-// chat-bubble timeline. User prompts on the right, assistant turns on
-// the left (with their tool calls expanded), tool results in muted
-// frames. The conversation is the agent's reasoning trail — operators
-// scan it to understand WHY the agent did what it did, not just WHAT
-// it did (the step timeline already covers that).
-//
-// Robustness: the artifact's content is JSON parsed inline. If the
-// JSON is corrupt we render an inline error and continue rendering
-// the rest of the run UI — losing the conversation viewer is much
-// better than crashing the whole page.
-function AgentConversationView({
-  raw,
-  canRetryFromTurn = false,
-  busy = false,
-  onRetryFromTurn,
-  steps = [],
-  streamTurnCosts,
-}: {
-  raw: string;
-  canRetryFromTurn?: boolean;
-  busy?: boolean;
-  // reason is collected via the inline modal before the call is made.
-  onRetryFromTurn?: (turn: number, reason: string) => void;
-  steps?: TaskStepRecord[];
-  // streamTurnCosts is a turn → µUSD map pushed by the SSE stream.
-  // Used to fill in costs missing from the model-step output_summary
-  // path. Optional — the steps path is the primary source.
-  streamTurnCosts?: Map<number, number>;
-}) {
-  // pendingRetryTurn is set when the operator clicks "↻ retry from here"
-  // on a bubble. The inline modal collects an optional reason, then
-  // fires onRetryFromTurn(turn, reason) on confirm.
-  const [pendingRetryTurn, setPendingRetryTurn] = useState<number | null>(null);
-
-  let messages: AgentConversationMessage[] = [];
-  try {
-    const parsed = JSON.parse(raw);
-    if (Array.isArray(parsed)) messages = parsed as AgentConversationMessage[];
-  } catch {
-    return (
-      <div style={{ padding: "10px 16px", borderBottom: "1px solid var(--border)", fontSize: 11, color: "var(--red)", fontFamily: "var(--font-mono)" }}>
-        Could not parse agent conversation artifact (invalid JSON).
-      </div>
-    );
-  }
-  if (messages.length === 0) return null;
-  const visibleMessages = messages.filter(isVisibleConversationMessage);
-  if (visibleMessages.length === 0) return null;
-
-  // Compute per-message turn numbers up-front so each bubble can render
-  // its own "↻ retry from turn N" affordance. Only assistant messages
-  // get a turn number — user/tool/system messages get 0 and won't show
-  // the button. Counting in a single pass here keeps the bubble itself
-  // O(1) at render time.
-  let assistantSeen = 0;
-  const turnByIndex: number[] = visibleMessages.map(m => {
-    if (m.role === "assistant") {
-      assistantSeen++;
-      return assistantSeen;
-    }
-    return 0;
-  });
-
-  // Build turn → cost map by walking model-kind steps in step.index
-  // order. The Nth model step corresponds to the Nth assistant turn,
-  // so we just zip them. Steps whose OutputSummary lacks the cost
-  // field (older runs, or resumed-after-approval steps that didn't
-  // re-call the LLM) map to undefined; we fall back to the SSE-
-  // streamed per-turn cost (when available) so the figure still
-  // shows up; otherwise the bubble renders nothing rather than a
-  // misleading "$0.000".
-  const costByTurn = buildTurnCostMap(steps);
-  if (streamTurnCosts) {
-    for (const [turn, micros] of streamTurnCosts) {
-      if (!costByTurn.has(turn) && micros > 0) {
-        costByTurn.set(turn, micros);
-      }
-    }
-  }
-
-  return (
-    <>
-      <div style={{ padding: "12px 16px", borderBottom: "1px solid var(--border)", display: "flex", flexDirection: "column", gap: 8 }}>
-        <div className="kicker" style={{ marginBottom: 4 }}>
-          Agent conversation · {visibleMessages.length} message{visibleMessages.length === 1 ? "" : "s"}
-        </div>
-        {visibleMessages.map((m, i) => (
-          <ConversationBubble
-            key={i}
-            message={m}
-            turn={turnByIndex[i]}
-            turnCostMicros={turnByIndex[i] > 0 ? costByTurn.get(turnByIndex[i]) : undefined}
-            canRetryFromTurn={canRetryFromTurn}
-            busy={busy || pendingRetryTurn !== null}
-            onRetryFromTurn={canRetryFromTurn ? (turn) => setPendingRetryTurn(turn) : undefined}
-          />
-        ))}
-      </div>
-      {pendingRetryTurn !== null && (
-        <RetryFromTurnModal
-          turn={pendingRetryTurn}
-          busy={busy}
-          onConfirm={(reason) => {
-            onRetryFromTurn?.(pendingRetryTurn, reason);
-            setPendingRetryTurn(null);
-          }}
-          onClose={() => setPendingRetryTurn(null)}
-        />
-      )}
-    </>
-  );
-}
-
-function isVisibleConversationMessage(message: AgentConversationMessage): boolean {
-  return message.role === "user" || message.role === "assistant" || message.role === "tool";
-}
-
-// RetryFromTurnModal collects an optional reason before submitting the
-// retry-from-turn request. The reason is stored in the run.resumed_from_event event
-// and shown in the run timeline so operators can annotate why they
-// branched from a particular turn.
-function RetryFromTurnModal({
-  turn,
-  busy,
-  onConfirm,
-  onClose,
-}: {
-  turn: number;
-  busy: boolean;
-  onConfirm: (reason: string) => void;
-  onClose: () => void;
-}) {
-  const [reason, setReason] = useState("");
-  return (
-    <Modal
-      title={`Retry from turn ${turn}`}
-      onClose={onClose}
-      width={440}
-      footer={
-        <div style={{ display: "flex", gap: 8, justifyContent: "flex-end" }}>
-          <button className="btn btn-ghost" onClick={onClose} disabled={busy}>
-            Cancel
-          </button>
-          <button
-            className="btn btn-primary"
-            disabled={busy}
-            onClick={() => onConfirm(reason.trim())}
-          >
-            {busy ? "Working…" : "Retry"}
-          </button>
-        </div>
-      }
-    >
-      <div style={{ display: "flex", flexDirection: "column", gap: 10 }}>
-        <p style={{ margin: 0, fontSize: 13, color: "var(--t1)", lineHeight: 1.5 }}>
-          A new run will be created with the conversation truncated to just before
-          turn {turn}'s assistant message. The prior steps and file state are preserved.
-        </p>
-        <label style={{ display: "flex", flexDirection: "column", gap: 4 }}>
-          <span style={{ fontSize: 11, fontWeight: 500, color: "var(--t2)", textTransform: "uppercase", letterSpacing: "0.05em" }}>
-            Reason <span style={{ fontWeight: 400, color: "var(--t3)" }}>(optional)</span>
-          </span>
-          <textarea
-            autoFocus
-            rows={2}
-            value={reason}
-            onChange={e => setReason(e.target.value)}
-            placeholder="Why are you branching from this turn?"
-            style={{
-              resize: "vertical", fontFamily: "var(--font-sans)", fontSize: 13,
-              color: "var(--t0)", background: "var(--bg2)",
-              border: "1px solid var(--border)", borderRadius: "var(--radius-sm)",
-              padding: "6px 8px", lineHeight: 1.5, width: "100%", boxSizing: "border-box",
-            }}
-          />
-        </label>
-      </div>
-    </Modal>
-  );
-}
-
-// buildTurnCostMap walks `steps` in step.index order, picks out the
-// model-kind ones, and pairs them with their turn ordinal. The agent
-// loop emits exactly one model step per turn (resumed-after-approval
-// turns use a different ToolName but still count as model steps), so
-// "first model step" = turn 1, "second" = turn 2, etc. Returns a map
-// keyed by turn number with the cost in µUSD; turns whose step has no
-// cost in OutputSummary are simply absent from the map.
-function buildTurnCostMap(steps: TaskStepRecord[]): Map<number, number> {
-  const sorted = [...steps].sort((a, b) => (a.index ?? 0) - (b.index ?? 0));
-  const out = new Map<number, number>();
-  let turn = 0;
-  for (const step of sorted) {
-    if (step.kind !== "model") continue;
-    turn++;
-    const summary = step.output_summary as Record<string, unknown> | undefined;
-    if (!summary) continue;
-    const raw = summary["cost_micros_usd"];
-    if (typeof raw === "number" && raw > 0) {
-      out.set(turn, raw);
-    }
-  }
-  return out;
-}
-
-function ConversationBubble({
-  message,
-  turn,
-  turnCostMicros,
-  canRetryFromTurn = false,
-  busy = false,
-  onRetryFromTurn,
-}: {
-  message: AgentConversationMessage;
-  turn?: number;
-  turnCostMicros?: number;
-  canRetryFromTurn?: boolean;
-  busy?: boolean;
-  onRetryFromTurn?: (turn: number) => void;
-}) {
-  if (message.role === "user") {
-    return (
-      <div style={{ display: "flex", justifyContent: "flex-end" }}>
-        <div style={{
-          maxWidth: "80%", padding: "8px 12px",
-          background: "var(--teal-bg)", border: "1px solid var(--teal-border)",
-          borderRadius: "var(--radius)", color: "var(--t0)", fontSize: 13, lineHeight: 1.5,
-          whiteSpace: "pre-wrap", wordBreak: "break-word",
-        }}>
-          {message.content || ""}
-        </div>
-      </div>
-    );
-  }
-  if (message.role === "tool") {
-    // Tool results are typically formatted "status=…\n--- stdout ---\n…"
-    // — render as a code block with monospace + scroll for long outputs.
-    const callRef = message.tool_call_id ? ` · ${message.tool_call_id.slice(0, 12)}` : "";
-    return (
-      <div style={{ display: "flex", justifyContent: "flex-start" }}>
-        <div style={{
-          maxWidth: "90%", padding: "6px 10px",
-          background: "var(--bg2)", border: "1px solid var(--border)",
-          borderRadius: "var(--radius-sm)", fontSize: 11,
-          fontFamily: "var(--font-mono)", color: "var(--t1)",
-        }}>
-          <div className="kicker" style={{ marginBottom: 4 }}>
-            tool result{callRef}
-          </div>
-          <pre style={{ margin: 0, whiteSpace: "pre-wrap", wordBreak: "break-word", maxHeight: 200, overflowY: "auto", color: "var(--t1)" }}>
-            {message.content || ""}
-          </pre>
-        </div>
-      </div>
-    );
-  }
-  // assistant — content + any tool calls
-  const showRetry = canRetryFromTurn && !!turn && turn > 0 && !!onRetryFromTurn;
-  const showCost = typeof turnCostMicros === "number" && turnCostMicros > 0;
-  return (
-    <div style={{ display: "flex", justifyContent: "flex-start", flexDirection: "column", gap: 6, alignItems: "stretch" }}>
-      <div className="kicker" style={{ display: "flex", alignItems: "center", gap: 8 }}>
-        <span>turn {turn || "?"}</span>
-        {showCost && (
-          <span
-            title={`LLM cost for turn ${turn}: ${formatMicrosUSD(turnCostMicros!)}`}
-            style={{ color: "var(--t3)", fontFamily: "var(--font-mono)" }}
-          >
-            · {formatMicrosUSD(turnCostMicros!)}
-          </span>
-        )}
-        {showRetry && (
-          <button
-            type="button"
-            className="btn btn-ghost btn-sm"
-            disabled={busy}
-            onClick={() => onRetryFromTurn?.(turn!)}
-            title={`Re-run from turn ${turn} with the prior conversation preserved`}
-            style={{ fontFamily: "var(--font-mono)", fontSize: 10, padding: "2px 6px" }}
-          >
-            ↻ retry from here
-          </button>
-        )}
-      </div>
-      {message.content && (
-        <div style={{
-          alignSelf: "flex-start", maxWidth: "80%", padding: "8px 12px",
-          background: "var(--bg3)", border: "1px solid var(--border)",
-          borderRadius: "var(--radius)", color: "var(--t0)", fontSize: 13, lineHeight: 1.5,
-          whiteSpace: "pre-wrap", wordBreak: "break-word",
-        }}>
-          {message.content}
-        </div>
-      )}
-      {message.tool_calls && message.tool_calls.length > 0 && (
-        <div style={{ display: "flex", flexDirection: "column", gap: 4 }}>
-          {message.tool_calls.map((tc, i) => (
-            <ToolCallChip key={tc.id || i} call={tc} />
-          ))}
-        </div>
-      )}
-    </div>
-  );
-}
-
-function ToolCallChip({ call }: { call: NonNullable<AgentConversationMessage["tool_calls"]>[number] }) {
-  // Pretty-print the JSON arguments when possible — collapsed to a
-  // single line for compactness, with a click-to-expand affordance.
-  const argsText = (() => {
-    if (!call.function?.arguments) return "";
-    try {
-      const parsed = JSON.parse(call.function.arguments);
-      return JSON.stringify(parsed);
-    } catch {
-      return call.function.arguments;
-    }
-  })();
-  return (
-    <div style={{
-      alignSelf: "flex-start", maxWidth: "90%",
-      padding: "6px 10px", background: "var(--bg2)",
-      border: "1px solid var(--teal-border)", borderRadius: "var(--radius-sm)",
-      fontFamily: "var(--font-mono)", fontSize: 11, color: "var(--t1)",
-    }}>
-      <span style={{ color: "var(--teal)", fontWeight: 500 }}>→ {call.function?.name || "(unknown)"}</span>
-      {argsText && (
-        <>
-          <span style={{ color: "var(--t3)" }}> </span>
-          <span title={argsText} style={{ color: "var(--t2)", overflow: "hidden", textOverflow: "ellipsis", whiteSpace: "nowrap", maxWidth: "100%", display: "inline-block", verticalAlign: "bottom" }}>
-            {argsText.length > 200 ? argsText.slice(0, 200) + "…" : argsText}
-          </span>
-        </>
-      )}
-    </div>
-  );
-}
